Added support for gwt-devmode, modified task depends.
gwt-compile was previously added as a dependent of prepare-webapp, but
prepare-webapp must be run before gwt-devmode and clearly one does not want to
run a long, painful gwt-compile before every gwt-devmode invocation.

Instead now gwt-compile and gwt-devmode depend on prepare-webapp, which makes
more sense, and package-war depends on gwt-compile. The one caveat is that if
you use run-jetty, it does not automatically call gwt-compile for you. However,
this is not totally crazy, since you might want to run-jetty once, and then
gwt-compile multiple times if you are crazy enough to be edit/compile/debugging
via gwt-compile instead of gwt-devmode.

I also changed gwt-compile to place its build results directly into
target/webapp because that is how the GWT compiler works in every other build
system I've seen it wired into, and it makes it more consistent with
gwt-devmode which necessarily must be told about target/webapp.

Lastly, added gwt-set-module which is used to tell gwt-devmode which module to
use, in the event that a project has multiple modules. In the absence of
specific configuration, the first module found is used.